[Refactor/#23]: 아이콘 import 방식 변경
