+++ public/assets/icons/index.ts
@@ -1,0 +1,15 @@
+export { default as AddIcon } from './add.svg?react';
+export { default as AiIcon } from './ai.svg?react';
+export { default as CalendarIcon } from './calendar.svg?react';
+export { default as CheckIcon } from './check.svg?react';
+export { default as CloseIcon } from './close.svg?react';
+export { default as DownArrowIcon } from './down-arrow.svg?react';
+export { default as EditIcon } from './edit.svg?react';
+export { default as EyeOnIcon } from './eye.svg?react';
+export { default as EyeOffIcon } from './eye2.svg?react';
+export { default as InfoIcon } from './info.svg?react';
+export { default as GlassesIcon } from './reading-glasses.svg?react';
+export { default as RemoveIcon } from './remove.svg?react';
+export { default as ReviewIcon } from './review.svg?react';
+export { default as CartIcon } from './shoppingbag.svg?react';
+export { default as WavingIcon } from './waving.svg?react';
