+++ public/assets/images/index.ts
@@ -1,0 +1,2 @@
+export { default as Poster } from './poster.svg?react';
+export { default as Logo } from './logo.svg?react';
