--- src/constants/icons.ts
@@ -1,7 +1,0 @@
-const PATH = '/assets/icons';
-
-export const ICONS = {
-  eye_on: `${PATH}/eye.svg`,
-  eye_off: `${PATH}/eye2.svg`,
-  down_arrow: `${PATH}/down-arrow.svg`,
-} as const;
